# apps/backend/main/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/profile/avatar": {
            "patch": {
                "description": "This endpoint uploads the user avatar image to the image storage service and updates the user profile information in the Auth0 Management API using the user's ID from the session. The ID is obtained from the session and used to query the user data from the external identity provider (Auth0). The user must be authenticated, and a valid session must exist. The request body must contain a valid image file.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Upload user avatar",
                "parameters": [
                    {
                        "type": "file",
                        "description": "User avatar image file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X coordinate of the crop area",
                        "name": "x",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Y coordinate of the crop area",
                        "name": "y",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Width of the crop area",
                        "name": "width",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Height of the crop area",
                        "name": "height",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile data",
                        "schema": {
                            "$ref": "#/definitions/postgres.User"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized if session is missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "404": {
                        "description": "User not found in Auth0 database",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error if request to Auth0 fails",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    }
                }
            }
        },
...
        "/workspaces/{workspace_id}": {
            "get": {
                "description": "Get workspace",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workspaces"
                ],
                "summary": "Get workspace",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Comma separated list of fields to include. Available fields: pages, owner",
                        "name": "include",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/postgres.Workspace"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized if session is missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "403": {
                        "description": "No access to workspace",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    }
                }
            },
            "put": {
                "description": "Update workspace by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workspaces"
                ],
                "summary": "Update workspace by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Workspace object that needs to be updated",
                        "name": "workspace",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/workspacesRouter.updateWorkspaceBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/postgres.Workspace"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized if session is missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "403": {
                        "description": "No access to workspace or no access to update workspace",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    }
                }
            }
        },
...
        "/workspaces/{workspace_id}/pages": {
            "get": {
                "description": "Get page list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pages"
                ],
                "summary": "Get page list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Format of the page list",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/postgres.Page"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized if session is missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "403": {
                        "description": "No access to workspace",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new page",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pages"
                ],
                "summary": "Create a new page",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Page object that needs to be created",
                        "name": "page",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pagesRouter.createPageBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/postgres.Page"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized if session is missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "403": {
                        "description": "No access to workspace or no access to create page",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "errorHandlers.Error": {
            "type": "object",
            "properties": {
                "details": {},
                "error": {
                    "type": "string"
                },
                "errorCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
        "pagesRouter.createPageBody": {
            "type": "object",
            "required": [
                "title",
                "type"
            ],
            "properties": {
                "parentId": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "enum": [
                        "text",
                        "board",
                        "group"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/postgres.PageType"
                        }
                    ]
                }
            }
        },
...
        "postgres.Page": {
            "type": "object",
            "properties": {
                "childPages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/postgres.Page"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/postgres.User"
                },
                "parentPage": {
                    "$ref": "#/definitions/postgres.Page"
                },
                "role": {
                    "$ref": "#/definitions/postgres.UserRole"
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/postgres.Task"
                    }
                },
                "text": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/postgres.PageType"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "postgres.PageType": {
            "type": "string",
            "enum": [
                "board",
                "text",
                "group"
            ],
            "x-enum-varnames": [
                "PageTypeBoard",
                "PageTypeText",
                "PageTypeGroup"
            ]
        },
        "postgres.Task": {
            "type": "object",
            "properties": {
                "assignee": {
                    "$ref": "#/definitions/postgres.User"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dueDate": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "page": {
                    "$ref": "#/definitions/postgres.Page"
                },
                "reporter": {
                    "$ref": "#/definitions/postgres.User"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "postgres.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "emailVerified": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "lastLogin": {
                    "type": "string"
                },
                "lastPasswordReset": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "postgres.UserRole": {
            "type": "string",
            "enum": [
                "owner",
                "admin",
                "member",
                "commentator",
                "guest"
            ],
            "x-enum-varnames": [
                "UserRoleOwner",
                "UserRoleAdmin",
                "UserRoleMember",
                "UserRoleCommentator",
                "UserRoleGuest"
            ]
        },
        "postgres.Workspace": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/postgres.User"
                },
                "pages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/postgres.Page"
                    }
                },
                "role": {
                    "$ref": "#/definitions/postgres.UserRole"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "workspacesRouter.updateWorkspaceBody": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        }
    }
}